% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pandoc.R
\name{add_rstudio_pandoc_to_rprofile_site}
\alias{add_rstudio_pandoc_to_rprofile_site}
\title{Add `RSTUDIO_PANDOC` environment variable to Rprofile.site}
\usage{
add_rstudio_pandoc_to_rprofile_site(app_dir)
}
\arguments{
\item{app_dir}{The application directory}
}
\description{
Adds the `RSTUDIO_PANDOC` environment variable to Rprofile.site so rmarkdown/htmlwidgets/etc know where to find pandoc
Assumes pandoc is in the R folder
}
