% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyBackgroundJS.R
\name{addFunctionToBackgroundJs}
\alias{addFunctionToBackgroundJs}
\title{Modify background.js to include the call to the shiny app}
\usage{
addFunctionToBackgroundJs(backgroundjsPath, packageName, functionName)
}
\arguments{
\item{backgroundjsPath}{path to the final background.js, not the one in inst/...}

\item{packageName}{package name, will be used for namespacing- (e.g. 'dplyr' in 'dplyr::filter()')}

\item{functionName}{function that runs your shiny app - (e.g. 'filter' in 'dplyr::filter()')}
}
\value{
none, side effect
}
\description{
Modify background.js to include the call to the shiny app
}
