% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appInstallR.R
\name{appInstallR}
\alias{appInstallR}
\title{TODO: Install user app on machine}
\usage{
appInstallR(repo)
}
\arguments{
\item{repo}{use repo or package with electron installer}
}
\value{
NA
}
\description{
TODO: Install user app on machine
}
