% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElectronPackager.R
\name{buildElectronDependencies}
\alias{buildElectronDependencies}
\title{Install electron-packager and electron via npm}
\usage{
buildElectronDependencies(app_root_path, node_path = NULL,
  npm_path = NULL, electricShine_nodejs = file.path(system.file(package
  = "electricShine"), "nodejs"))
}
\arguments{
\item{app_root_path}{path of new electron app}

\item{node_path}{path to node.exe}

\item{npm_path}{path to npm-cli.js}

\item{buildPath}{path of nodejs folder}
}
\value{
nothing
}
\description{
Install electron-packager and electron via npm
}
