% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunction.R
\name{buildPackage}
\alias{buildPackage}
\title{Meta-function}
\usage{
buildPackage(appName = "My_Package",
  description = "My Electron application", productName = "productName",
  version = NULL, installTo = NULL, MRANdate = Sys.Date() - 3,
  functionName = NULL, githubRepo = NULL, localPath = NULL)
}
\arguments{
\item{appName}{electron app name}

\item{description}{electron app description}

\item{productName}{necessary?}

\item{version}{semantic version of your app, as character (not numeric!)}

\item{installTo}{path to create installer, preferably points to an empty directory}

\item{MRANdate}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{functionName}{the function name in your package that starts the shiny app}

\item{githubRepo}{GitHub username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP')}

\item{localPath}{path to local shiny-app package}
}
\value{
Nothing
}
\description{
Meta-function
}
