% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_utils.R
\name{check_conda}
\alias{check_conda}
\title{Check that conda works}
\usage{
check_conda(conda_path)
}
\arguments{
\item{conda_path}{path to conda executable}
}
\value{

}
\description{
Check that conda works
}
