% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_utils.R
\name{conda_create_env}
\alias{conda_create_env}
\title{Create a new conda environment}
\usage{
conda_create_env(conda_top_dir, conda_env)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}
}
\value{

}
\description{
Create a new conda environment
}
