% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install.R
\name{conda_install}
\alias{conda_install}
\title{Install a conda package}
\usage{
conda_install(
  conda_top_dir,
  conda_env = "eshine",
  conda_package,
  conda_repo = NULL
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{conda_package}{conda package to be installed}

\item{conda_repo}{conda repository package will be installed from}
}
\value{

}
\description{
Install a conda package
}
