% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install.R
\name{conda_install_nodejs}
\alias{conda_install_nodejs}
\title{Instal python from conda}
\usage{
conda_install_nodejs(
  conda_top_dir,
  conda_repo = "conda-forge",
  conda_env = "eshine-nodejs",
  nodejs_version = "15.13.0"
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_repo}{conda repository package will be installed from}

\item{conda_env}{conda environment name}

\item{nodejs_version}{semantic version of nodejs to install (e.g. :15.13.0")}
}
\value{

}
\description{
Instal python from conda
}
