% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_pack.R
\name{conda_install_pack}
\alias{conda_install_pack}
\title{Install conda-pack}
\usage{
conda_install_pack(conda_top_dir, conda_env = NULL, conda_repo = "conda-forge")
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{conda_repo}{conda repository package will be installed from}
}
\value{

}
\description{
Install conda-pack
}
