% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install.R
\name{conda_install_python}
\alias{conda_install_python}
\title{Instal python from conda}
\usage{
conda_install_python(
  conda_top_dir,
  conda_repo = "conda-forge",
  conda_env = "eshine",
  python_version = "3.7"
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_repo}{conda repository package will be installed from}

\item{conda_env}{conda environment name}

\item{python_version}{semantic version of python to install (e.g. "3.7")}
}
\value{

}
\description{
Instal python from conda
}
