% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install.R
\name{conda_install_r}
\alias{conda_install_r}
\title{Install R from conda}
\usage{
conda_install_r(
  conda_top_dir,
  conda_env = "eshine",
  conda_repo = "conda-forge",
  r_version = NULL
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{conda_repo}{conda repository package will be installed from}

\item{r_version}{semantic version of R to install (e.g. "4.0")}
}
\value{

}
\description{
Install R from conda
}
