% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_pack.R
\name{conda_pack}
\alias{conda_pack}
\title{Package a conda environment with conda-pack}
\usage{
conda_pack(conda_top_dir, conda_env = "eshine", outdir)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{outdir}{directory path where conda-pack 'tar.gz' should be written to}
}
\value{

}
\description{
Package a conda environment with conda-pack
}
