% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_doc.R
\name{conda_params}
\alias{conda_params}
\title{Conda parameter documentation}
\usage{
conda_params(conda_top_dir, conda_env, conda_package, conda_repo)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{conda_package}{conda package to be installed}

\item{conda_repo}{conda repository package will be installed from}

\item{r_package_repo}{cran-like repo to install R packages from}
}
\value{

}
\description{
Conda parameter documentation
}
