% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_update.R
\name{conda_update}
\alias{conda_update}
\title{Update conda version}
\usage{
conda_update(conda_top_dir)
}
\arguments{
\item{conda_top_dir}{path to miniconda}
}
\value{

}
\description{
Update conda version
}
