% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_mran_url.R
\name{construct_mran_url}
\alias{construct_mran_url}
\title{Construct MRAN url if snapshot date provided}
\usage{
construct_mran_url(mran_date = NULL, cran_like_url = NULL)
}
\arguments{
\item{mran_date}{MRAN snapshot date (e.g. "2019-08-05")}

\item{cran_like_url}{url to CRAN-like repository}
}
\value{
CRAN-like url
}
\description{
Construct MRAN url if snapshot date provided
}
\examples{

construct_mran_url(mran_date = "2019-08-05")

construct_mran_url(cran_like_url = "https://cloud.r-project.org")

}
