% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{copy_electricshine_package}
\alias{copy_electricshine_package}
\title{Copy {electricShine} package to an isolated folder.
   This is necessary to avoid dependency-install issues}
\usage{
copy_electricshine_package()
}
\value{
path of new {electricShine}-only library
}
\description{
Copy {electricShine} package to an isolated folder.
   This is necessary to avoid dependency-install issues
}
