% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R, R/install_user_app_new.R
\name{copy_remotes_package}
\alias{copy_remotes_package}
\title{Copy {remotes} package to an isolated folder.
   This is necessary to avoid dependency-install issues}
\usage{
copy_remotes_package(
  new_path = file.path(tempdir(), "electricShine", "templib")
)

copy_remotes_package(
  new_path = file.path(tempdir(), "electricShine", "templib")
)
}
\value{
path of new {remotes}-only library

path of new {remotes}-only library
}
\description{
Copy {remotes} package to an isolated folder.
   This is necessary to avoid dependency-install issues

Copy {remotes} package to an isolated folder.
   This is necessary to avoid dependency-install issues
}
