% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_template.R
\name{copy_template}
\alias{copy_template}
\title{Copy Electron boilerplate into app_root_path}
\usage{
copy_template(app_root_path)
}
\arguments{
\item{app_root_path}{directory of new Electron app}
}
\value{
none
}
\description{
Copy Electron boilerplate into app_root_path
}
