% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAppFile.R
\name{create_app_file}
\alias{create_app_file}
\title{Create the app.R file that will launch the app}
\usage{
create_app_file(packageName, path)
}
\arguments{
\item{packageName}{name of the shiny app package}

\item{path}{path of the electricShine folder}
}
\value{
nothing
}
\description{
Create the app.R file that will launch the app
}
