% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryTemplate.R
\name{create_app_R}
\alias{create_app_R}
\title{Create app.R file}
\usage{
create_app_R(packageName, path)
}
\arguments{
\item{packageName}{name of package}

\item{path}{path of where to write app.R}
}
\value{
nothing, writes app.R to path provided and provides feedback if succcessful or not
}
\description{
Create app.R file
}
