% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_build_directory.R
\name{create_build_directory}
\alias{create_build_directory}
\title{Create a directory for creating the new app and copy template of files}
\usage{
create_build_directory(name, description, productName, version, appPath,
  functionName, ...)
}
\arguments{
\item{name}{name of app}

\item{description}{short description of app}

\item{productName}{product name}

\item{version}{version number: see https://semver.org/ for details on how to use version numbers}

\item{appPath}{path to new electron app top directory}

\item{functionName}{the function name in your package that starts the shiny app}

\item{...}{pass optional arguments to electricShine::create_package_json()}
}
\value{
nothing, creates a directory
}
\description{
Create a directory for creating the new app and copy template of files
}
