% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryTemplate.R
\name{create_directory_template}
\alias{create_directory_template}
\title{Create a directory for creating the new app and copy template of files}
\usage{
create_directory_template(path)
}
\arguments{
\item{path}{path to create the new app}
}
\value{
nothing
}
\description{
Create a directory for creating the new app and copy template of files
}
