% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Create_Folder}
\alias{Create_Folder}
\title{Title  Create an output folder}
\usage{
Create_Folder(path, name)
}
\arguments{
\item{path}{path where output folder gonna be}

\item{name}{name of the output folder about to be created}
}
\value{
if folder name already exists, show error and do nothing;
        if folder name doesn't exist, create a new file
}
\description{
Title  Create an output folder
}
\examples{

}
