% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMainJs.R
\name{create_main_js}
\alias{create_main_js}
\title{Create main.js file}
\usage{
create_main_js(path)
}
\arguments{
\item{path}{path of where to write main.js}
}
\value{
nothing, writes main.js to path provided and provides feedback if succcessful or not
}
\description{
Create main.js file
}
