% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageJson.R
\name{create_package_json}
\alias{create_package_json}
\title{Create the package.json file for npm}
\usage{
create_package_json(name, description, productName, version = NULL,
  appPath, iconPath, repository = "", author = "", license = "",
  electronVersion = "^4.0.7", electronPackagerVersion = "^13.1.1")
}
\arguments{
\item{name}{name of app}

\item{description}{short description of app}

\item{productName}{product name}

\item{version}{version number (as string, not numeric): see https://semver.org/ for details on how to use version numbers}

\item{appPath}{path to created app}

\item{iconPath}{path to icon within created app}

\item{repository}{purely for info- does the shiny app live in a repository (e.g. GitHub)}

\item{author}{author of the app}

\item{license}{license of the App. Not the full license, only the title (e.g. MIT, or GPLv3)}

\item{electronVersion}{version of electron that should be downloaded from npm and used to create app}

\item{electronPackagerVersion}{version of electronPackager that should be downloaded from npm and used to create app}
}
\description{
Create the package.json file for npm
}
