% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm_package_lock.R
\name{create_package_lock_json}
\alias{create_package_lock_json}
\title{Create package-lock.json for npm}
\usage{
create_package_lock_json(path)
}
\arguments{
\item{path}{path to write package-lock.json for npm}
}
\value{
nothing, writes package-lock.json to path provided and provides feedback if succcessful or not
}
\description{
Create package-lock.json for npm
}
