% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryTemplate.R
\name{create_renderer_js}
\alias{create_renderer_js}
\title{Create renderer.js file}
\usage{
create_renderer_js(path)
}
\arguments{
\item{path}{path of where to write renderer.js}
}
\value{
nothing, writes renderer.js to path provided and provides feedback if succcessful or not
}
\description{
Create renderer.js file
}
