% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_arch}
\alias{.check_arch}
\title{Check if compatible architecture}
\usage{
.check_arch(arch = base::version$arch[[1]])
}
\arguments{
\item{arch}{only used for unit testing}
}
\value{
Stops if unsupported architecture
}
\description{
Check if compatible architecture
}
