% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_build_path_exists}
\alias{.check_build_path_exists}
\title{Check whether build path exists}
\usage{
.check_build_path_exists(build_path)
}
\arguments{
\item{build_path}{build_path}
}
\value{
stops
}
\description{
Check whether build path exists
}
