% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_node_version_format}
\alias{.check_node_version_format}
\title{Add "v" to nodejs version if needed}
\usage{
.check_node_version_format(node_version)
}
\arguments{
\item{node_version}{nodejs version stringsss}
}
\value{
nodejs version in format "v10.16.0" not "10.16.0"
}
\description{
Add "v" to nodejs version if needed
}
