% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_nodejs_npm.R
\name{.check_node_works}
\alias{.check_node_works}
\title{Check if Node works}
\usage{
.check_node_works(node_top_dir, expected_version)
}
\arguments{
\item{node_top_dir}{directory containing nodejs app/exe}

\item{expected_version}{expected version of node}
}
\value{
path of nodejs executable if found and functional, otherwise: FALSE
}
\description{
Check if Node works
}
