% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_nodejs_npm.R
\name{.check_npm_works}
\alias{.check_npm_works}
\title{Check if npm works}
\usage{
.check_npm_works(node_top_dir)
}
\arguments{
\item{node_top_dir}{directory containing npmjs app/exe}
}
\value{
path of npm executable if found and functional, otherwise: FALSE
}
\description{
Check if npm works
}
