% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_package_provided}
\alias{.check_package_provided}
\title{Check package paths}
\usage{
.check_package_provided(arguments)
}
\arguments{
\item{arguments}{arguments to check}
}
\value{
stops
}
\description{
Doesn't check whether it works, only for conflicts in arguments
}
