% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_repo_set}
\alias{.check_repo_set}
\title{Check that repo is set}
\usage{
.check_repo_set(arguments)
}
\arguments{
\item{arguments}{arguments to check}
}
\value{
stops if neither cran-like url or mran date is set
}
\description{
Doesn't check whether it works, only that it is not-null
}
