% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r.R
\name{.download_r}
\alias{.download_r}
\title{Download R installer given its url}
\usage{
.download_r(d_url)
}
\arguments{
\item{d_url}{download file from url}
}
\value{
Path to R.exe installer
}
\description{
Download R installer given its url
}
