% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{.install_mac_r}
\alias{.install_mac_r}
\title{Download and untar mac R into app folder}
\usage{
.install_mac_r(app_root_path, mac_url)
}
\arguments{
\item{app_root_path}{top level of new electricShine app build}

\item{mac_url}{url for mac R language download}
}
\value{
NA
}
\description{
Download and untar mac R into app folder
}
