% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{.install_win_r}
\alias{.install_win_r}
\title{Install R for Windows at given path}
\usage{
.install_win_r(win_installer_path, app_root_path)
}
\arguments{
\item{win_installer_path}{path of Windows R installer}

\item{app_root_path}{top level of new electricShine app build}
}
\value{
NA, installs R to path
}
\description{
Install R for Windows at given path
}
