% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{.prompt_install_nodejs}
\alias{.prompt_install_nodejs}
\title{Prompt whether nodejs can be installed or not}
\usage{
.prompt_install_nodejs(nodejs_path)
}
\arguments{
\item{nodejs_path}{nodejs install path}
}
\value{
TRUE/FALSE
}
\description{
Prompt whether nodejs can be installed or not
}
