% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{.prompt_install_r}
\alias{.prompt_install_r}
\title{Prompt whether R can be installed or not}
\usage{
.prompt_install_r(app_root_path)
}
\arguments{
\item{app_root_path}{project path}
}
\value{
TRUE/FALSE
}
\description{
Prompt whether R can be installed or not
}
