% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_download_install.R
\name{download_miniconda3}
\alias{download_miniconda3}
\title{Download miniconda install script}
\usage{
download_miniconda3(os, minconda_version = "latest", minconda_tempdir = NULL)
}
\arguments{
\item{os}{this development operating system ("win", "mac", or "lin")}

\item{minconda_version}{miniconda version}

\item{minconda_tempdir}{where should the install script be saved?}
}
\value{

}
\description{
Download miniconda install script
}
