% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLocal.R
\name{electricShine}
\alias{electricShine}
\title{Meta-function}
\usage{
electricShine(path, date = "2019-01-01", package)
}
\arguments{
\item{path}{path to create installer}

\item{date}{date for MRAN}

\item{package}{github username/repo}
}
\value{
Nothing
}
\description{
Meta-function
}
