% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_function.R
\name{electrify}
\alias{electrify}
\title{Meta-function}
\usage{
electrify(
  app_name = "example_app",
  product_name = "product_name",
  short_description = NULL,
  semantic_version = NULL,
  build_path = "/Users/chase/Documents/eshine_temp",
  mran_date = NULL,
  cran_like_url = "https://cran.r-project.org",
  function_name = NULL,
  shiny_package_location = "github",
  shiny_package_path = "chasemc/demoapp",
  shiny_package_install_opts = NULL,
  run_build = TRUE,
  nodejs_path = file.path(system.file(package = "electricShine"), "nodejs"),
  nodejs_version = "v12.16.2",
  permission = FALSE,
 
    mac_url = "https://mac.r-project.org/el-capitan/R-3.6-branch/R-3.6-branch-el-capitan-sa-x86_64.tar.gz",
  conda_env = "ehsine",
  r_version = NULL,
  ...
)
}
\arguments{
\item{app_name}{This will be the name of the executable. It's a uniform type identifier (UTI)
that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters.
see https://www.electron.build/configuration/configuration
and https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070}

\item{product_name}{String - allows you to specify a product name for your executable which
contains spaces and other special characters not allowed in the name property.
https://www.electron.build/configuration/configuration}

\item{short_description}{short app description}

\item{semantic_version}{semantic version of your app, as character (not numeric!);
See https://semver.org/ for more info on semantic versioning.}

\item{build_path}{Path where the build files will be created, preferably points to an empty directory.
Must not contain a folder with the name as what you put for electrify(app_name).}

\item{mran_date}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{cran_like_url}{url to cran-like repository}

\item{function_name}{the function name in your package that starts the shiny app}

\item{run_build}{logical, whether to start the build process, helpful if you want to modify anthying before building}

\item{nodejs_path}{path to nodejs}

\item{nodejs_version}{nodejs version to install}

\item{permission}{automatically grant permission to install nodejs and R}

\item{mac_url}{url to mac OS tar.gz}

\item{package_location}{one of c("local", "github", "gitlab", "bitbucket")}

\item{package_path}{path to local shiny-app package. If 'git_host' isn't used, use host/repo (e.g. chasemc/electricShine)}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}
}
\description{
Meta-function
}
