% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_utils.R
\name{find_conda_program}
\alias{find_conda_program}
\title{Return filepath of conda, given miniconda directory path}
\usage{
find_conda_program(conda_top_dir)
}
\arguments{
\item{conda_top_dir}{path to miniconda}
}
\value{

}
\description{
Return filepath of conda, given miniconda directory path
}
