% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodejs.R
\name{find_nodejs}
\alias{find_nodejs}
\title{Find nodejs and npm executables}
\usage{
find_nodejs(electricShine_nodejs = file.path(system.file(package =
  "electricShine"), "nodejs"))
}
\arguments{
\item{electricShine_nodejs}{electricShine pacakge location or location of nodejs directory}
}
\value{
list of two executable paths, one for nodejs and one for npm
}
\description{
Find nodejs and npm executables
}
