% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodejs.R
\name{get_nodejs}
\alias{get_nodejs}
\title{Install Node.js}
\usage{
get_nodejs(node_url = "https://nodejs.org/dist",
  nodejs_path = file.path(system.file(package = "electricShine"),
  "nodejs"), force_install = FALSE, node_version = "v10.16.0")
}
\arguments{
\item{node_url}{path to node.js.org}

\item{nodejs_path}{where node.js will be installed to}

\item{node_version}{version of node.js (eg "v10.15.1")}

\item{node_url}{should node.js be installed if it's already present?}
}
\value{
list of 2: "nodePath": path to node.exe; "npmPath": path to npm-cli.js
}
\description{
Install Node.js
}
