% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodejs.R
\name{getElectron}
\alias{getElectron}
\title{Download Electron}
\usage{
getElectron(nodePath = NULL, npmPath = NULL,
  installTo = file.path(system.file(package = "electricShine"),
  "nodejs"))
}
\arguments{
\item{nodePath}{path to node.exe}

\item{npmPath}{path to npm-cli.js}

\item{installTo}{path of nodejs folder}
}
\value{
nothing
}
\description{
Download Electron
}
