% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_download_install.R
\name{install_miniconda3}
\alias{install_miniconda3}
\title{Install miniconda3}
\usage{
install_miniconda3(
  miniconda_install_script_path,
  miniconda_installation_path = NULL
)
}
\arguments{
\item{miniconda_install_script_path}{file path of the miniconda installer script}

\item{miniconda_installation_path}{directory path where conda will be installed}
}
\value{

}
\description{
Install miniconda3
}
