% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_nodejs.R
\name{install_nodejs}
\alias{install_nodejs}
\title{Install Node.js}
\usage{
install_nodejs(
  node_url = "https://nodejs.org/dist",
  nodejs_path = file.path(system.file(package = "electricShine"), "nodejs"),
  force_install = FALSE,
  nodejs_version = "v15.5.0",
  permission_to_install = FALSE
)
}
\arguments{
\item{node_url}{path to node.js.org}

\item{nodejs_path}{parent folder of node.exe (~nodejs_path/node.exe)}

\item{force_install}{try and install even if nodejs exists  (TRUE/FALSE)}

\item{nodejs_version}{version of node.js (eg "v10.16.0")}

\item{permission_to_install}{whether to skip .prompt_install_nodejs()}
}
\value{
installed or checked nodejs_path
}
\description{
Install Node.js
}
