% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r.R
\name{install_r}
\alias{install_r}
\title{Install R from MRAN date into electricShine folder}
\usage{
install_r(cran_like_url = NULL, app_root_path,
  mac_url = "https://mac.r-project.org/el-capitan/R-3.6-branch/R-3.6-branch-el-capitan-sa-x86_64.tar.gz")
}
\arguments{
\item{cran_like_url}{CRAN-like url e.g. https://cran.r-project.org/bin/windows/base}

\item{app_root_path}{path to current electricShine app build}

\item{mac_url}{mac R installer url}
}
\description{
Install R from MRAN date into electricShine folder
}
