% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install_r_package.R
\name{install_r_packages}
\alias{install_r_packages}
\title{Install R package within conda}
\usage{
install_r_packages(
  r_package_name,
  r_package_repo = "https://cran.r-project.org",
  conda_top_dir,
  conda_env = "eshine"
)
}
\arguments{
\item{r_package_name}{name of package to install}

\item{r_package_repo}{cran-like repo to install R packages from}

\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}
}
\value{

}
\description{
Install R package within conda
}
