% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install_r_package.R
\name{install_r_remotes}
\alias{install_r_remotes}
\title{Install {remotes} R package}
\usage{
install_r_remotes(
  conda_top_dir,
  r_package_repo = "https://cran.r-project.org",
  conda_env = "eshine"
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{r_package_repo}{cran-like repo to install R packages from}

\item{conda_env}{conda environment name}
}
\value{

}
\description{
Install {remotes} R package
}
