% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_install_r_package_with_remotes.R
\name{install_remote_package}
\alias{install_remote_package}
\title{Currently broken, installs to local r need to fix env-vars like in previous release}
\usage{
install_remote_package(
  conda_top_dir,
  conda_env = "eshine",
  repo_location = "github",
  repo = "chasemc/demoapp",
  dependencies_repo = "https://cran.r-project.org",
  package_install_opts = NULL
)
}
\arguments{
\item{conda_top_dir}{path to miniconda}

\item{conda_env}{conda environment name}

\item{repo_location}{{remotes} package function, one of c("github", "gitlab", "bitbucket", "local")}

\item{repo}{e.g. if repo_location is github: "chasemc/demoApp" ; if repo_location is local: "C:/Users/chase/demoApp"}

\item{dependencies_repo}{cran-like repo to install R package dependencies from}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}
}
\value{

}
\description{
Currently broken, installs to local r need to fix env-vars like in previous release
}
