% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installUserApp.R
\name{install_user_app}
\alias{install_user_app}
\title{Title}
\usage{
install_user_app(package, path, date)
}
\arguments{
\item{package}{github username/repo    eg tidyverse/ggplot2}

\item{path}{path to electricShine folder}

\item{date}{MRAN date from which to download packages from}
}
\value{
nothing
}
\description{
Title
}
