% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{installR}
\alias{installR}
\title{Install R from MRAN date into electricShine folder}
\usage{
installR(date, exe = "R-[0-9.]+.+-win\\\\.exe", path)
}
\arguments{
\item{date}{MRAN date from which to install R}

\item{exe}{find exe name}

\item{path}{path to electricShine directory}
}
\value{
nothing
}
\description{
Install R from MRAN date into electricShine folder
}
