% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r_portable_mac.R
\name{modify_mac_r}
\alias{modify_mac_r}
\title{Change fixed paths to make R portable on Mac}
\usage{
modify_mac_r(r_executable_path)
}
\arguments{
\item{r_executable_path}{path to ~bin/R}
}
\value{
NA, modifies file
}
\description{
Change fixed paths to make R portable on Mac
}
