% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_helpers.R
\name{pull_from_description}
\alias{pull_from_description}
\title{Tells electricShine to pull an item from the description, and possibly run through a post processing function
This usually isn't done immediately, to use the exact package version that will be provided with the app
If you want the result returned immediately (or use a different DESCRIPTION file), you can specify `description_object`, which should be the return of [desc::desc]}
\usage{
pull_from_description(
  fieldname,
  default = "",
  postprocessing_function = NULL,
  description_object
)
}
\arguments{
\item{fieldname}{The field to pull from the description}

\item{default}{The default value, if the entry is missing from description}

\item{postprocessing_function}{A function that takes the description output and returns a string to put in package.json}

\item{description_object}{An option}
}
\description{
Tells electricShine to pull an item from the description, and possibly run through a post processing function
This usually isn't done immediately, to use the exact package version that will be provided with the app
If you want the result returned immediately (or use a different DESCRIPTION file), you can specify `description_object`, which should be the return of [desc::desc]
}
\examples{
\dontrun{
electrify(
   app_name = pull_from_description("Package"),
   semantic_version = pull_from_description("Version", r_ver_to_semver),
   short_description = pull_from_description("Description")
)
}
}
