% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_helpers.R
\name{r_authors_string_to_node_first_author}
\alias{r_authors_string_to_node_first_author}
\title{Parse an R authors string to something fit for package.json
npm only does one author, and rest in AUHORS.md or contributor, we only do author here}
\usage{
r_authors_string_to_node_first_author(r_authors_string)
}
\arguments{
\item{r_authors_string}{A scalar R author string. This should contain valid R code and will be evaluated as such.}
}
\value{
A package.json author string
}
\description{
Parse an R authors string to something fit for package.json
npm only does one author, and rest in AUHORS.md or contributor, we only do author here
}
\examples{
\dontrun{
r_authors_string_to_node_first_author(desc::desc(package = "electricShine")$get_field("Authors@R"))
r_authors_string_to_node_first_author(desc::desc(package = "desc")$get_field("Authors@R"))
}
}
