% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_helpers.R
\name{r_ver_to_semver}
\alias{r_ver_to_semver}
\title{Transforms an R development version (4 dots) to a semver development version (3 dots and a dash)}
\usage{
r_ver_to_semver(r_ver)
}
\arguments{
\item{r_ver}{A string containing an R package version, possibly a development version}
}
\value{
A semver version
}
\description{
This function does not validate the version, be sure to make sure it's valid
The version is expected to be formatted as recommended in [R packages](https://r-pkgs.org/lifecycle.html#version) and as created by `usethis::use_version()`
}
\examples{
r_ver_to_semver("1.0.0") # returns unmodified
r_ver_to_semver("0.0.1.9001") # dev version dashed
r_ver_to_semver("1") # Not an appropriate R version, no dots added, doesn't result in a valid semver
}
