% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electron_build_release.R
\name{run_build}
\alias{run_build}
\title{Create an electron-builder release}
\usage{
run_build(node_path = NULL, npm_path = NULL, app_path,
  node = file.path(system.file(package = "electricShine"), "nodejs"))
}
\arguments{
\item{node_path}{option to specifiy the path to the node.js directory if already installed}

\item{npm_path}{option to specifiy the path to the npm module directory if already installed}

\item{app_path}{path to new electron app top directory}

\item{node}{folder where electricShine installs/looks for node an npm if not given in node_path/npm_path}
}
\value{
nothing, used for side-effects
}
\description{
Create an electron-builder release
}
