% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electron_build_release.R
\name{run_build_release}
\alias{run_build_release}
\title{Create an electron-builder release}
\usage{
run_build_release(
  nodejs_path = file.path(system.file(package = "electricShine"), "nodejs"),
  app_path,
  nodejs_version
)
}
\arguments{
\item{nodejs_path}{parent folder of node.exe (~nodejs_path/node.exe)}

\item{app_path}{path to new electron app top directory}

\item{nodejs_version}{for checking if nodejs is functional}
}
\value{
nothing, used for side-effects
}
\description{
Create an electron-builder release
}
