% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShinyApp.R
\name{run_Shiny}
\alias{run_Shiny}
\title{Create app.R file that starts the shiny app}
\usage{
run_Shiny(packageName, path, functionName)
}
\arguments{
\item{packageName}{name of package}

\item{path}{path of where to write app.R}

\item{functionName}{function that starts the shiny app

functionName should be the name of the function that startsthe package's shiny app.
e.g. is you have the function myPackage::start_shiny(), provide "start_shiny"}
}
\value{
nothing, writes app.R to path provided and provides feedback if succcessful or not
}
\description{
Create app.R file that starts the shiny app
}
