% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{system_install_pkgs}
\alias{system_install_pkgs}
\title{Run package installation using the newly-installed R}
\usage{
system_install_pkgs(rscript_path)
}
\arguments{
\item{rscript_path}{path to newly-installed R's executable}
}
\value{
nothing
}
\description{
Run package installation using the newly-installed R
}
