% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app_new.R
\name{system_install_pkgs_new}
\alias{system_install_pkgs_new}
\title{Run package installation using the newly-installed R}
\usage{
system_install_pkgs_new(rscript_path, arguments)
}
\arguments{
\item{rscript_path}{path to newly-installed R's executable}

\item{arguments}{arguments to the install script}
}
\value{
nothing
}
\description{
Run package installation using the newly-installed R
}
