% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimR.R
\name{trim_r}
\alias{trim_r}
\title{Remove html and pdf files from R installation}
\usage{
trim_r(pathToR, only64)
}
\arguments{
\item{pathToR}{path to the copied R installation}

\item{only64}{if TRUE, remove 32-bit dlls; if FALSE do not remove 32-bit dlls}
}
\value{
nothing
}
\description{
Remove html and pdf files from R installation
}
