% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_text}
\alias{write_text}
\title{Write to a file}
\usage{
write_text(text, name, path)
}
\arguments{
\item{name}{name of file to be created, including extension}

\item{path}{path where file should be written to}
}
\value{
message saying whether file was created or not
}
\description{
Write to a file
}
