% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunction.R
\name{buildElectricApp}
\alias{buildElectricApp}
\title{Meta-function}
\usage{
buildElectricApp(app_name = NULL, product_name = "product_name",
  short_description = NULL, semantic_version = NULL,
  build_path = NULL, mran_date = NULL, cran_like_url = NULL,
  function_name = NULL, github_repo = NULL, local_package = NULL,
  package_name = NULL, build = TRUE, nodejs_path = NULL,
  nodejs_version = NULL)
}
\arguments{
\item{app_name}{electron app name}

\item{product_name}{necessary?}

\item{short_description}{short app description}

\item{semantic_version}{semantic version of your app, as character (not numeric!)}

\item{build_path}{path to create installer, preferably points to an empty directory}

\item{mran_date}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{function_name}{the function name in your package that starts the shiny app}

\item{github_repo}{GitHub username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP')}

\item{local_package}{path to local shiny-app package}

\item{package_name}{name of your R package, because a git repo doesn't have to be the same as the package name}

\item{build}{logical, whether to start the build process, helpful if you want to modify anthying before building}
}
\value{
Nothing
}
\description{
Meta-function
}
