% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_function.R
\name{electrify}
\alias{electrify}
\title{Meta-function}
\usage{
electrify(app_name = NULL, product_name = "product_name",
  short_description = NULL, semantic_version = NULL,
  build_path = NULL, mran_date = NULL, cran_like_url = NULL,
  function_name = NULL, git_host = NULL, git_repo = NULL,
  local_package_path = NULL, package_install_opts = NULL,
  run_build = TRUE, nodejs_path = file.path(system.file(package =
  "electricShine"), "nodejs"), nodejs_version = "v10.16.0",
  permission = FALSE)
}
\arguments{
\item{app_name}{This will be the name of the executable. It's a uniform type identifier (UTI)
that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters.
see https://www.electron.build/configuration/configuration 
and https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070}

\item{product_name}{String - allows you to specify a product name for your executable which
contains spaces and other special characters not allowed in the name property.
https://www.electron.build/configuration/configuration}

\item{short_description}{short app description}

\item{semantic_version}{semantic version of your app, as character (not numeric!);
See https://semver.org/ for more info on semantic versioning.}

\item{build_path}{Path where th bbuild files will be created, preferably points to an empty directory.
Must not contain a folder with the name as what you put for electrify(app_name).}

\item{mran_date}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{cran_like_url}{url to cran-like repository}

\item{function_name}{the function name in your package that starts the shiny app}

\item{git_host}{one of c("github", "gitlab", "bitbucket")}

\item{git_repo}{GitHub/Bitbucket/GitLab username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP'). 
Can also use notation for commits/branch (i.e. "chasemc/demoapp@d81fff0).}

\item{local_package_path}{path to local shiny-app package, if 'git_package' isn't used}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{run_build}{logical, whether to start the build process, helpful if you want to modify anthying before building}

\item{nodejs_path}{path to nodejs}

\item{nodejs_version}{nodejs version to install}

\item{permission}{automatically grant permission to install nodejs and R}
}
\description{
Meta-function
}
