% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunction.R
\name{buildElectricApp}
\alias{buildElectricApp}
\title{Meta-function}
\usage{
buildElectricApp(appName = "My_Package", packageName = NULL,
  description = "My Electron application", productName = "productName",
  semanticVersion = NULL, installTo = NULL, MRANdate = Sys.Date() -
  3, functionName = NULL, githubRepo = NULL, localPath = NULL,
  only64 = FALSE)
}
\arguments{
\item{appName}{electron app name}

\item{productName}{necessary?}

\item{semanticVersion}{semantic version of your app, as character (not numeric!)}

\item{installTo}{path to create installer, preferably points to an empty directory}

\item{MRANdate}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{functionName}{the function name in your package that starts the shiny app}

\item{githubRepo}{GitHub username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP')}

\item{localPath}{path to local shiny-app package}

\item{only64}{if TRUE, remove 32-bit dlls; if FALSE do not remove 32-bit dlls}
}
\value{
Nothing
}
\description{
Meta-function
}
