% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElectronPackager.R
\name{buildElectronDependencies}
\alias{buildElectronDependencies}
\title{Install electron-packager and electron via npm}
\usage{
buildElectronDependencies(appPath, nodePath = NULL, npmPath = NULL,
  installTo = file.path(system.file(package = "electricShine"),
  "nodejs"))
}
\arguments{
\item{appPath}{path of new electron app}

\item{nodePath}{path to node.exe}

\item{npmPath}{path to npm-cli.js}

\item{installTo}{path of nodejs folder}
}
\value{
nothing
}
\description{
Install electron-packager and electron via npm
}
