% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunction.R
\name{buildPackage}
\alias{buildPackage}
\title{Meta-function}
\usage{
buildPackage(name = "My_Package",
  description = "My Electron application", productName = "productName",
  version = NULL, path = NULL, date = "2019-01-01", package = NULL)
}
\arguments{
\item{path}{path to create installer}

\item{date}{date for MRAN}

\item{package}{github username/repo}
}
\value{
Nothing
}
\description{
Meta-function
}
