% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{copy_remotes_package}
\alias{copy_remotes_package}
\title{Copy {remotes} package to an isolated folder.
   This is necessary to avoid dependency-install issues}
\usage{
copy_remotes_package()
}
\value{
path of new {remotes}-only library
}
\description{
Copy {remotes} package to an isolated folder.
   This is necessary to avoid dependency-install issues
}
