% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_template.R
\name{copy_template}
\alias{copy_template}
\title{Copy Electron boilerplate into appPath}
\usage{
copy_template(appPath)
}
\arguments{
\item{appPath}{directory of new Electron app}
}
\value{
none
}
\description{
Copy Electron boilerplate into appPath
}
