% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folder}
\alias{create_folder}
\title{Create an output folder}
\usage{
create_folder(app_root_path)
}
\arguments{
\item{app_root_path}{path where output folder gonna be}
}
\value{
if folder name already exists, show error and do nothing;
}
\description{
Create an output folder
}
