% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryTemplate.R
\name{create_package_json}
\alias{create_package_json}
\title{Create the package.json file for npm}
\usage{
create_package_json(name, description, productName, version = NULL, path)
}
\arguments{
\item{name}{name of app}

\item{description}{short description of app}

\item{productName}{product name}

\item{version}{version number: see https://semver.org/ for details on how to use version numbers}

\item{path}{path to create the new app}
}
\description{
Create the package.json file for npm
}
