% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package_json.R
\name{create_package_json}
\alias{create_package_json}
\title{Create the package.json file for npm}
\usage{
create_package_json(
  app_name = "MyApp",
  description = "description",
  semantic_version = "0.0.0",
  app_root_path = NULL,
  repository = "",
  author = "",
  copyright_year = "",
  copyright_name = "",
  website = "",
  license = "",
  deps = NULL
)
}
\arguments{
\item{app_name}{name of your app. This is what end-users will see/call an app}

\item{description}{short description of app}

\item{semantic_version}{semantic version of app see https://semver.org/ for more information on versioning}

\item{app_root_path}{app_root_path to where package.json will be written}

\item{repository}{purely for info- does the shiny app live in a repository (e.g. GitHub)}

\item{author}{author of the app}

\item{copyright_year}{year of copyright}

\item{copyright_name}{copyright-holder's name}

\item{website}{website of app or company}

\item{license}{license of the App. Not the full license, only the title (e.g. MIT, or GPLv3)}

\item{deps}{is to allow testing with testthat}
}
\value{
outputs package.json file with user-input modifications
}
\description{
Create the package.json file for npm
}
