% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_package_provided}
\alias{.check_package_provided}
\title{Check package paths}
\usage{
.check_package_provided(git_host, git_repo, local_package_path)
}
\arguments{
\item{git_host}{git_host}

\item{git_repo}{git_repo}

\item{local_package_path}{local_package_path}
}
\value{
stops
}
\description{
Doesn't check whether it works, only for conflicts in arguments
}
