% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_fast.R
\name{.check_repo_set}
\alias{.check_repo_set}
\title{Check that a repo for packages/R was set}
\usage{
.check_repo_set(cran_like_url = NULL, mran_date = NULL)
}
\arguments{
\item{cran_like_url}{cran_like_url}

\item{mran_date}{mran_date}
}
\description{
Check that a repo for packages/R was set
}
