% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{.download_r}
\alias{.download_r}
\title{Download R installer for windows to tempdir()}
\usage{
.download_r(d_url)
}
\arguments{
\item{url}{url to download R from}
}
\value{
Path to R.exe installer
}
\description{
Download R installer for windows to tempdir()
}
