% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{.find_win_exe_url}
\alias{.find_win_exe_url}
\title{Find Windows R installer URL from MRAN snapshot}
\usage{
.find_win_exe_url(cran_like_url = NULL, mran_date = NULL)
}
\arguments{
\item{mran_date}{MRAN date (yyyy-mm-dd) to download from}
}
\value{
url
}
\description{
Find Windows R installer URL from MRAN snapshot
}
