% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_function.R
\name{electrify}
\alias{electrify}
\title{Meta-function}
\usage{
electrify(
  app_name,
  product_name = NULL,
  short_description = NULL,
  semantic_version = NULL,
  build_path = NULL,
  mran_date = NULL,
  cran_like_url = NULL,
  function_name = NULL,
  git_host = NULL,
  git_repo = NULL,
  local_package_path = NULL,
  package_install_opts = NULL,
  dependency_install_opts = NULL,
  include_remotes = FALSE,
  run_build = TRUE,
  nodejs_path = NULL,
  nodejs_version = "v12.16.2",
  permission = FALSE,
  mac_url = NULL,
  r_bitness = c("x64", "i386"),
  app_args = "",
  pandoc_version = NULL,
  r_version = "latest",
  copyright = "",
  author = NULL,
  website = NULL,
  license = NULL,
  keywords = "electron-app",
  include_rtools = FALSE,
  platforms = list(),
  deps = NULL
)
}
\arguments{
\item{app_name}{This will be the name of the executable. It's a uniform type identifier (UTI)
that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters.
see https://www.electron.build/configuration/configuration
and https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070}

\item{product_name}{String - allows you to specify a product name for your executable which
contains spaces and other special characters not allowed in the name property.
https://www.electron.build/configuration/configuration}

\item{short_description}{short app description}

\item{semantic_version}{semantic version of your app, as character (not numeric!);
See https://semver.org/ for more info on semantic versioning.}

\item{build_path}{Path where the build files will be created, preferably points to an empty directory.
Must not contain a folder with the name as what you put for electrify(app_name).}

\item{mran_date}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{cran_like_url}{url to cran-like repository}

\item{function_name}{the function name in your package that starts the shiny app}

\item{git_host}{one of c("github", "gitlab", "bitbucket")}

\item{git_repo}{GitHub/Bitbucket/GitLab username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP').
Can also use notation for commits/branch (i.e. "chasemc/demoapp@d81fff0).}

\item{local_package_path}{path to local shiny-app package, if 'git_package' isn't used}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{dependency_install_opts}{optional arguments to remotes::install_deps, if NULL then remotes will not pre-install dependencies}

\item{include_remotes}{Whether or not to copy remotes to the local package path}

\item{run_build}{logical, whether to start the build process, helpful if you want to modify anthying before building}

\item{nodejs_path}{path to nodejs}

\item{nodejs_version}{nodejs version to install}

\item{permission}{automatically grant permission to install nodejs and R}

\item{mac_url}{url to mac OS tar.gz}

\item{r_bitness}{The bitness of the R installation you want to use (i386 or x64)}

\item{app_args}{Quoted arguments to the Shiny app, for example `"options = list(port = 1010)"` to set a fixed port.
These are inserted into a JS string then quoted and executed through shell, so keep in mind complex quoting}

\item{pandoc_version}{pandoc version to install, as in `as.character(rmarkdown::pandoc_version())`. If null, pandoc won't be installed}

\item{r_version}{R version to install. If set to 'latest', the latest version will be installed}

\item{author}{author of the app}

\item{website}{website of app or company}

\item{license}{license of the App. Not the full license, only the title (e.g. MIT, or GPLv3)}

\item{keywords}{Keywords describing the app}

\item{include_rtools}{Whether or not to include RTools. Requires installr to install. Will increase app size substantially}

\item{platforms}{Named list containing build instructions for `win`, `mac` and `linux`}

\item{deps}{Named list containing Node.js dependencies}
}
\description{
Meta-function
}
