% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodejs.R
\name{get_nodejs}
\alias{get_nodejs}
\title{Install Node.js}
\usage{
get_nodejs(nodeUrl = "https://nodejs.org/dist",
  buildPath = file.path(system.file(package = "electricShine"),
  "nodejs"), force = FALSE, nodeVersion = "v10.16.0")
}
\arguments{
\item{nodeUrl}{path to node.js.org}

\item{buildPath}{where node.js will be installed to}

\item{force}{should node.js be installed if it's already present?}

\item{nodeVersion}{version of node.js (eg "v10.15.1")}
}
\value{
list of 2: "nodePath": path to node.exe; "npmPath": path to npm-cli.js
}
\description{
Install Node.js
}
