% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installR.R
\name{install_r}
\alias{install_r}
\title{Install R from MRAN date into electricShine folder}
\usage{
install_r(mran_date = as.character(Sys.Date() - 3), app_root_path,
  mac_url = "https://mac.r-project.org/el-capitan/R-3.6-branch/R-3.6-branch-el-capitan-sa-x86_64.tar.gz")
}
\arguments{
\item{mran_date}{MRAN date from which to install R}

\item{app_root_path}{path to current electricShine app build}
}
\value{
NA, installs/downloads R to a given path
}
\description{
Install R from MRAN date into electricShine folder
}
