% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installUserApp.R
\name{install_user_app}
\alias{install_user_app}
\title{Install your shiny app package and its dependencies to the R installation's library}
\usage{
install_user_app(appPath = NULL, MRANdate = NULL, githubRepo = NULL,
  localPath = NULL, ...)
}
\arguments{
\item{appPath}{path to the Electron app's build folder}

\item{MRANdate}{MRAN date from which to download packages from}

\item{githubRepo}{GitHub username/repo of your the shiny-app package  (e.g. 'chasemc/demoAPP')}

\item{localPath}{path to local shiny-app package}
}
\value{
nothing
}
\description{
Install your shiny app package and its dependencies to the R installation's library
}
