% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installUserApp.R
\name{install_user_app}
\alias{install_user_app}
\title{Install your shiny app package and its dependencies to the R installation's library}
\usage{
install_user_app(app_root_path = NULL, mran_date = NULL,
  github_repo = NULL, local_path = NULL)
}
\arguments{
\item{app_root_path}{path to the Electron app's build folder}

\item{mran_date}{MRAN date from which to download packages from}

\item{github_repo}{GitHub username/repo of your the shiny-app package  (e.g. 'chasemc/demoAPP')}

\item{local_path}{path to local shiny-app package}
}
\value{
nothing
}
\description{
Install your shiny app package and its dependencies to the R installation's library
}
