% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_backgroundjs.R
\name{modify_background_js}
\alias{modify_background_js}
\title{Modify background.js to include the call to the shiny app}
\usage{
modify_background_js(
  background_js_path,
  my_package_name,
  function_name,
  r_path,
  r_bitness = "x64",
  app_args
)
}
\arguments{
\item{background_js_path}{path to the final background.js, not the one in inst/...}

\item{my_package_name}{package name, will be used for namespacing- (e.g. 'dplyr' in 'dplyr::filter()')}

\item{function_name}{function that runs your shiny app - (e.g. 'filter' in 'dplyr::filter()')}

\item{r_path}{path from "r_lang" folder to the R/Rscript executable}

\item{r_bitness}{bitness of R installation, only applicable before R 4.2.0}

\item{app_args}{arguments to the function that runs your shiny app}
}
\value{
none, side effect
}
\description{
Modify background.js to include the call to the shiny app
}
