% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryTemplate.R
\name{setup_directory}
\alias{setup_directory}
\title{Create a directory for creating the new app and copy template of files}
\usage{
setup_directory(name, description, productName, version, appPath)
}
\arguments{
\item{name}{name of app}

\item{description}{short description of app}

\item{productName}{product name}

\item{version}{version number: see https://semver.org/ for details on how to use version numbers}

\item{path}{path to create the new app}
}
\value{
nothing, creates a directory
}
\description{
Create a directory for creating the new app and copy template of files
}
