% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_r.R
\name{trim_r}
\alias{trim_r}
\title{Remove html and pdf files from R installation}
\usage{
trim_r(app_root_path)
}
\arguments{
\item{app_root_path}{path to the copied R installation}
}
\value{
nothing
}
\description{
Remove html and pdf files from R installation
}
