% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_text}
\alias{write_text}
\title{Write to a file}
\usage{
write_text(text, filename, path)
}
\arguments{
\item{text}{text to write}

\item{filename}{name of file to write to}

\item{path}{path where file should be written to}
}
\value{
message saying whether file was created or not
}
\description{
Write to a file
}
